backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.14
uvicorn[standard]==0.24.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.1.0  # For text rendering fallback

# Task queue
celery==5.3.4
redis==5.0.1

# AI/LLM
google-genai==0.3.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
